Fix the fft interpolation to have less copies and the correct precision.

+++ romancal/refpix/tests/reference_utils.py
@@ -288,8 +288,10 @@ def fft_interp(
                 / chanFrameData_Flat.size
             )
             chanFrameData_Flat = spfft.irfft(
-                fftResult * chanFrameData_Flat.size, workers=1
-            )
+                fftResult * chanFrameData_Flat.size,
+                workers=1
+                # )
+            ).astype(chanFrameData_Flat.dtype)
             # Return read only pixels
             chanFrameData_Flat[readOnlyPixelsIndices_Flat] = readOnlyPixelsValues_Flat
